requirements_dev.txt
# general install dependencies
pip>=18.0
wheel>=0.30.0

# glotaran setup dependencies
pyglotaran==0.1.0


# testing dependencies
tox>=3.0.0
pytest>=3.7.1
pluggy>=0.7

# code quality asurence
pre-commit>=2.6.0
